% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_bets}
\alias{compute_bets}
\title{Compute Binary Bets
Takes predictions and creates lagged -1, 0, 1 values based on sign.}
\usage{
compute_bets(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Binary Bets
Takes predictions and creates lagged -1, 0, 1 values based on sign.
}
