% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_bets_long}
\alias{compute_bets_long}
\title{Compute Binary Long Bets
Takes predictions and creates lagged 0, 1 values based on sign.}
\usage{
compute_bets_long(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Binary Long Bets
Takes predictions and creates lagged 0, 1 values based on sign.
}
