% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_bets_short}
\alias{compute_bets_short}
\title{Compute Binary Short Bets
Takes predictions and creates lagged -1, 0 values based on sign.}
\usage{
compute_bets_short(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Binary Short Bets
Takes predictions and creates lagged -1, 0 values based on sign.
}
