% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_hits_short}
\alias{compute_hits_short}
\title{Compute Short Hits
Returns 1 TRUE if binary bet == sign(log growth).}
\usage{
compute_hits_short(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Short Hits
Returns 1 TRUE if binary bet == sign(log growth).
}
