% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_return}
\alias{compute_return}
\title{Compute Return
Returns return based on bet.}
\usage{
compute_return(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Return
Returns return based on bet.
}
