% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_return_long}
\alias{compute_return_long}
\title{Compute Long Return
Returns long return based on bet.}
\usage{
compute_return_long(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Long Return
Returns long return based on bet.
}
