% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_return_short}
\alias{compute_return_short}
\title{Compute Short Return
Returns short return based on bet.}
\usage{
compute_return_short(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Short Return
Returns short return based on bet.
}
