% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalAR.R
\docType{package}
\name{conditionalAR}
\alias{conditionalAR}
\alias{conditionalAR-package}
\title{Conditionally Evaluate Rolling Performance of AR Models}
\description{
Conditionally Evaluate Rolling Performance of AR Models.
}
\author{
\strong{Maintainer}: Jason Muhlenkamp \email{jason.muhlenkamp@gmail.com}

}
