% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbhitrate.R
\name{lbhitrate}
\alias{lbhitrate}
\title{Compute Rolling Hit Rate From LBStart to LBEnd
Compute Rolling Hit Rate From LBStart to LBEnd.}
\usage{
lbhitrate(data, x, lbstart = 63, lbend = 0, position = "longshort")
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}

\item{lbstart}{The start of the lookback window.}

\item{lbend}{The end of the lookback window.}

\item{position}{Character for whether to return \cr
\itemize{
\item All "longshort" bets (-1, 0, 1 for all non-NA)
\item Only "long" bets (0, 1 for all non-NA)
\item Only "short" bets (-1, 0 for all non-NA)
\item "alwayslong" (1 for all non-NA obs)
\item "alwaysshort" (-1 for all non-NA obs)
}}
}
\description{
Compute Rolling Hit Rate From LBStart to LBEnd
Compute Rolling Hit Rate From LBStart to LBEnd.
}
\examples{
growth <- diff(log(stocks))
signal <- rollAR(growth, rep(1/9,9))

# Rolling 126-day long-short hitrate
lbhitrate(growth, signal, 126, 0, "longshort")
}
