% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-helpers.R
\name{pmax.xts}
\alias{pmax.xts}
\title{pmax for xts Objects}
\usage{
pmax.xts(x, ...)
}
\arguments{
\item{x}{An `xts` object.}

\item{...}{Arguments to be passed to `pmax`.}
}
\description{
Wrapper around `pmax` in order to apply to xts objects.
}
