% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-helpers.R
\name{pmin.xts}
\alias{pmin.xts}
\title{pmin for xts Objects}
\usage{
pmin.xts(x, ...)
}
\arguments{
\item{x}{An `xts` object.}

\item{...}{Arguments to be passed to `pmin`.}
}
\description{
Wrapper around `pmin` in order to apply to xts objects.
}
