% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollAR.R
\name{rollAR}
\alias{rollAR}
\title{Apply AR Coefficients}
\usage{
rollAR(x, coef)
}
\arguments{
\item{x}{An xts object}

\item{coef}{Numeric vector of AR coefficients.
First element corresponds to AR1.}
}
\description{
Apply AR coefficients to one or multiple columns of an xts object
in a rolling fashion.
}
