% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollARens.R
\name{rollARens}
\alias{rollARens}
\title{Apply an Ensemble of AR Coefficients}
\usage{
rollARens(x, coef)
}
\arguments{
\item{x}{An xts object}

\item{coef}{A list of numeric vectors of AR coefficients.
First element in each vector corresponds to AR1.}
}
\description{
Apply an ensemble of AR coefficients to one or multiple columns of an xts object
in a rolling fashion.
}
