% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Daily Closing Prices of Major European Stock Indices, 1991–1998}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 1860 rows and 4 columns.}
\usage{
stocks
}
\description{
datasets::EuStockMarkets in xts format.
}
\keyword{datasets}
