% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_betrate}
\alias{summary_betrate}
\title{Compute Bet Rate
Takes predictions and returns bet rate.}
\usage{
summary_betrate(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Bet Rate
Takes predictions and returns bet rate.
}
