% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_betrate_long}
\alias{summary_betrate_long}
\title{Compute Long Bet Rate
Takes signals and returns bet rate for long predictions.}
\usage{
summary_betrate_long(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Long Bet Rate
Takes signals and returns bet rate for long predictions.
}
