% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_betrate_short}
\alias{summary_betrate_short}
\title{Compute Short Bet Rate
Takes signals and returns bet rate for short predictions.}
\usage{
summary_betrate_short(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Short Bet Rate
Takes signals and returns bet rate for short predictions.
}
