% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_bets}
\alias{summary_bets}
\title{Compute Bets
Takes predictions and returns bet rate.}
\usage{
summary_bets(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Bets
Takes predictions and returns bet rate.
}
