% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_bets_long}
\alias{summary_bets_long}
\title{Compute Long Bets
Takes signals and returns bet count for long predictions.}
\usage{
summary_bets_long(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Long Bets
Takes signals and returns bet count for long predictions.
}
