% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_bets_short}
\alias{summary_bets_short}
\title{Compute Short Bets
Takes signals and returns bet count for short predictions.}
\usage{
summary_bets_short(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Short Bets
Takes signals and returns bet count for short predictions.
}
