% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_days}
\alias{summary_days}
\title{Compute Days
Takes predictions and returns N.}
\usage{
summary_days(x)
}
\arguments{
\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Days
Takes predictions and returns N.
}
