% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_downrate}
\alias{summary_downrate}
\title{Compute Down Rate
Takes predictions and growths and returns hit rate.}
\usage{
summary_downrate(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Down Rate
Takes predictions and growths and returns hit rate.
}
