% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_hitrate_short}
\alias{summary_hitrate_short}
\title{Compute Short Hit Rate
Takes predictions and growths and returns hit rate.}
\usage{
summary_hitrate_short(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Short Hit Rate
Takes predictions and growths and returns hit rate.
}
