% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rates.R
\name{summary_uprate}
\alias{summary_uprate}
\title{Compute Up Rate
Takes predictions and growths and returns hit rate.}
\usage{
summary_uprate(data, x)
}
\arguments{
\item{data}{An `xts` object containing log growths.}

\item{x}{An `xts` object containing predictions.}
}
\description{
Compute Up Rate
Takes predictions and growths and returns hit rate.
}
