% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transOMLS.R
\name{transOMLS}
\alias{transOMLS}
\title{Transform Raw Signals into OLOS, MLMS, etc.}
\usage{
transOMLS(signalcomb, x)
}
\arguments{
\item{signalcomb}{Character of either: olos, mlms, olml, osms.}

\item{x}{xts object containing the raw signals.}
}
\description{
Transform raw signals into OLOS, MLMS, etc.
}
