% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollAR-helpers.R
\name{weighted.sum}
\alias{weighted.sum}
\title{Weighted Arithmetic Sum}
\usage{
weighted.sum(x, w, ..., na.rm = FALSE)
}
\arguments{
\item{x}{an object containing the values whose weighted sum is to
be computed.}

\item{w}{a numerical vector of weights the same length as x giving
the weights to use for elements of x.}

\item{...}{arguments to be passed to or from methods.}

\item{na.rm}{a logical value indicating whether NA values in x
should be stripped before the computation proceeds.}
}
\description{
Compute a weighted sum.
}
