% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_bets}
\alias{compute_bets}
\title{Compute Binary Bets
Takes predictions and creates lagged -1, 0, 1 values based on sign.}
\usage{
compute_bets(x, position = c("longshort", "long", "short", "alwayslong",
  "alwaysshort"))
}
\arguments{
\item{x}{An `xts` object containing predictions.}

\item{position}{Character for whether to return \cr
\itemize{
\item All "longshort" bets (-1, 0, 1 for all non-NA)
\item Only "long" bets (0, 1 for all non-NA)
\item Only "short" bets (-1, 0 for all non-NA)
\item "alwayslong" (1 for all non-NA obs)
\item "alwaysshort" (-1 for all non-NA obs)
}}
}
\description{
Compute Binary Bets
Takes predictions and creates lagged -1, 0, 1 values based on sign.
}
