% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpapplyr.R
\name{lbapplyr}
\alias{lbapplyr}
\title{Apply Rolling Functions From LBStart to LBEnd
Wrapper around rollapplyr so that we can easily apply rolling functions
over lagged windows}
\usage{
lbapplyr(data, FUN, lbstart = 63, lbend = 1, ...)
}
\arguments{
\item{data}{An `xts` object that is passed to `rollapplyr`.}

\item{FUN}{The function to be applied.}

\item{lbstart}{The start of the lookback window.}

\item{lbend}{The end of the lookback window.}

\item{...}{Other arguments passed to rollapplyr.}
}
\description{
Apply Rolling Functions From LBStart to LBEnd
Wrapper around rollapplyr so that we can easily apply rolling functions
over lagged windows
}
\examples{
growth <- diff(log(stocks))
signal <- rollAR(growth, rep(1/9,9))
lbapplyr(abs(compute_bets(signal, "all")), sum, 63, 0)
lbapplyr(abs(compute_bets(signal, "long")), sum, 63, 0)
lbapplyr(abs(compute_bets(signal, "short")), sum, 63, 0)
lbapplyr(abs(compute_bets(signal, "n")), sum, 63, 0)
}
